Remove Azure Deployment
Not going to deploy this, due to the sensitive nature of the project. It is only really meant for local development, and focusing on deploying has impacted the release.

Now it is a backend tool for developers to use, having an Azure demo isn't worth it due to it being a personal product that I am not monetising with users.